requirements.txt
# Core dependencies
tensorflow>=2.8.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration and data validation
pyyaml>=6.0
pydantic>=1.9.0

# Development tools
pytest>=6.0
pytest-cov>=2.0
flake8>=3.9.0
mypy>=0.910
black>=21.5b2
isort>=5.9.0

# Jupyter and notebooks
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Testing and quality
coverage>=6.0.0
codecov>=2.1.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0
requests>=2.26.0

# Optional dependencies
torch>=1.10.0; platform_system != "Darwin"  # PyTorch for non-Mac systems
torch>=1.10.0; platform_system == "Darwin" and platform_machine == "arm64"  # PyTorch for M1 Macs

# Additional dependencies
pandapower>=2.8.0  # For power grid analysis
pvlib>=0.9.0       # For solar energy modeling
windpowerlib>=0.2.0  # For wind energy modeling 